Implement a new optimization of TD13 SIMD code
This change expands the TD13 SIMD implementation by adding a custom vectorized cosine to increase performance. It also obviates the need to build and install SLEEF for its cosine implementation.

This is the fastest I could make it without using floats and losing precision. I will add/update the documentation in future commits.

+++ include/crpropa/magneticField/TD13Field.h
@@ -96,7 +96,7 @@ private:
   std::vector<double> costheta;
   std::vector<double> beta;
 
-  std::vector<float> avx_data;
+  std::vector<double> avx_data;
   std::vector<double> Ak;
   std::vector<double> k;
